Renamed question research snapshot files
